writer: Simplify finish logic and reduce memory consumption
The logic in the original cdb library was to collect all the entries in
a single list, and then split them into separate tables during the
finish process. To do this, though, it ends up building an extra table
referencing each element.

This revised logic maintains separate lists of entries for each of the
256 hash tables. Then in finish, it just needs to build one hash table
at a time, which is a straight forward procedure.